refactor:Change of types of some attributes of memory item

+++ src/main/java/org/xenon/knowspace/dtos/MemoryItemDto.java
@@ -26,7 +26,7 @@ public class MemoryItemDto {
 
         private int interval;
 
-        private float easeFactor;
+        private double easeFactor;
 
         private int reviewCount;
